GameManager model create start

+++ GameServerClientExample/GameServer/Models/Player.cs
@@ -6,8 +6,8 @@ namespace GameServer.Models
         public long Id { get; set; }
         public string Name { get; set; }
         public long Score { get; set; }
-        public long PosX { get; set; }
-        public long PosY { get; set; }
+        public int PosX { get; set; }
+        public int PosY { get; set; }
 
         public Player()
         {
